Module InferBase.Memcached

val connect : unit -> unit

connect to a running memcached server -- only call this from processes which do not fork

val disconnect : unit -> unit

disconnect after having connected first

val start : unit -> unit

start a memcached daemon and set up an epilogue to kill it on exit -- only for top-level

module type Value = sig ... end

type to marshal, plus a unique label that will be colon-prepended to a key, roughly signifying a table

module type Server = sig ... end

module Make : functor (V : Value) -> Server with module Value = V