Module InferBase.ProcessPoolState

val in_child : bool InferStdlib.IStd.ref

Keep track of whether the current execution is in a child process

val update_status : (Mtime.t -> string -> unit) InferStdlib.IStd.ref

Ping the task bar whenever a new task is started with the start time and a description for the task

val get_pid : unit -> InferStdlib.IStd.Pid.t

val reset_pid : unit -> unit