Module InferBase__Multilinks

module F = Stdlib.Format

module L = InferBase.Logging

type t

In-memory representation of multilink files.

val add : t -> string -> unit

Add a link.

val create : unit -> t

Create a new multilink.

val multilink_file_name : string

Name of the multilink file. A multilink file is recognized by its file name.

val read : dir:string -> t option

Read a multilink file from disk.

val resolve : InferBase.DB.filename -> InferBase.DB.filename

Resolve a filename following multilinks. The cache is updated if a new multilinks file is read.

val reset_cache : unit -> unit

Reset the cache of multilink files

val write : t -> dir:string -> unit

Write a multilink file in the given directory