Library InferIR: Intermediate Representation

The Intermediate Representation, called SIL, is a format used by the back-end for analysis. It is produced by the front-ends.

The entry point of this library is the module: InferIR.

The main entry point is the intermediate language in InferIR.Sil.

The control flow graph module is InferIR.Cfg.

The type environment module is InferIR.Tenv.