Module Typ.Fieldname

type t

Names for fields of class/struct/union

val compare : t -> t -> int

val equal : t -> t -> bool

Equality for field names.

module Set : InferStdlib.IStd.Caml.Set.S with type Set.elt = t

Set for fieldnames

module Map : InferStdlib.IStd.Caml.Map.S with type Map.key = t

Map for fieldnames

module Clang : sig ... end

module Java : sig ... end

val to_string : t -> string

Convert a field name to a string.

val to_full_string : t -> string

val to_simplified_string : t -> string

Convert a fieldname to a simplified string with at most one-level path.

val to_flat_string : t -> string

Convert a fieldname to a flat string without path.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty print a field name.

val clang_get_qual_class : t -> QualifiedCppName.t option

get qualified classname of a field if it's coming from clang frontend. returns None otherwise