Module InferIR__Attributes

Module to manage the table of attributes.

type attributes_kind

val deserialize_attributes_kind : Sqlite3.Data.t -> attributes_kind

val store : proc_desc:InferIR.Procdesc.t option -> InferIR.ProcAttributes.t -> unit

Save .attr file for the procedure into the attributes database.

val load : InferIR.Typ.Procname.t -> InferIR.ProcAttributes.t option

Load the attributes for the procedure from the attributes database.

val load_defined : InferIR.Typ.Procname.t -> InferIR.ProcAttributes.t option

Load attributes for the procedure but only if is_defined is true

val find_file_capturing_procedure : InferIR.Typ.Procname.t -> (InferBase.SourceFile.t * [ `Include | `Source ]) option

Find the file where the procedure was captured, if a cfg for that file exists. Return also a boolean indicating whether the procedure is defined in an include file.

val pp_attributes_kind : Stdlib.Format.formatter -> attributes_kind -> unit