Module Odoc_document.Targets

Collect all the urls of pages defined by a model.

Roughly a simplified version of the normal process to convert a model into a document, only for extracting Urls. Used to determine the build targets.

let unit: Odoc_model.Lang.Compilation_unit.t => list( Url.Path.t );

let page: Odoc_model.Lang.Page.t => list( Url.Path.t );