Module Preamble_examples.Module

This comment is on the declaration of Module as an alias of Hidden__Module

This is the top comment declared in the module Hidden__module.

This is the second paragraph in the module Hidden__module.

This is the comment attached to the declaration of Hidden__Module

type t

This is a comment on type t