Module Odoc_model

module Lang : sig ... end

module Fold : sig ... end

This module allows to fold over odoc values. It is notably used to construct a search database of every relevant item. It appear to be very generic but in reality it is quite specialized to fold over searchable items, and not every kind of odoc value you could fold over.

module Comment : sig ... end

module Paths : sig ... end

module Names : sig ... end

Typed names for paths, identifiers, references and fragments.

module Predefined : sig ... end

module Root : sig ... end

A root can be seen as a unique representative of a odoc file.

module Error : sig ... end

module Location_ : sig ... end

module Compat : sig ... end

module Semantics : sig ... end