Module Odoc_model.Comment

module Path = Paths.Path ;

module Reference = Paths.Reference ;

module Identifier = Paths.Identifier ;

type with_location('a) = Location_.with_location ('a) ;

type style = [

| `Bold

| `Italic

| `Emphasis

| `Superscript

| `Subscript

];

type raw_markup_target = string;

type leaf_inline_element = [

| `Space

| `Word(string)

| `Code_span(string)

| `Raw_markup( ( raw_markup_target , string) )

];

type non_link_inline_element = [

| leaf_inline_element

| `Styled( (style , list( with_location ( non_link_inline_element ))) )

];

type link_content = list( with_location ( non_link_inline_element )) ;

type inline_element = [

| leaf_inline_element

| `Styled( (style , list( with_location ( inline_element ))) )

| `Reference( ( Reference.t , link_content ) )

| `Link( (string, link_content ) )

];

type paragraph = list( with_location ( inline_element )) ;

type module_reference = {

module_reference: Reference.Module.t ,

module_synopsis: option( paragraph ),

};

The {!modules: ...} markup. module_synopsis is initially None, it is resolved during linking.

type nestable_block_element = [

| `Paragraph( paragraph )

| `Code_block(string)

| `Verbatim(string)

| `Modules(list( module_reference ))

| `List( ([ `Unordered | `Ordered ], list( list( with_location ( nestable_block_element )))) )

];

type canonical_path = [

| `Root(string)

| `Dot( ( Path.Module.t , string) )

];

type tag = [

| `Author(string)

| `Deprecated(list( with_location ( nestable_block_element )))

| `Param( (string, list( with_location ( nestable_block_element ))) )

| `Raise( (string, list( with_location ( nestable_block_element ))) )

| `Return(list( with_location ( nestable_block_element )))

| `See( ( [ `Url | `File | `Document ] , string, list( with_location ( nestable_block_element ))) )

| `Since(string)

| `Before( (string, list( with_location ( nestable_block_element ))) )

| `Version(string)

| `Canonical( canonical_path )

| `Inline

| `Open

| `Closed

];

type heading_level = [

| `Title

| `Section

| `Subsection

| `Subsubsection

| `Paragraph

| `Subparagraph

];

type attached_block_element = [

| nestable_block_element

| `Tag(tag)

];

type block_element = [

| nestable_block_element

| `Heading( ( heading_level , Identifier.Label.t , link_content ) )

| `Tag(tag)

];

type docs = list( with_location ( block_element )) ;

type docs_or_stop = [

| `Docs( docs)

| `Stop

];

let synopsis: list( Location_.with_location ( [> `Paragraph('a) ] )) => option('a);

The synopsis is the first element of a comment if it is a paragraph. Otherwise, there is no synopsis.