Module Identifier.FunctorParameter

include IdSig with type t = t and type t_pv = t_pv

type t = t_pv id

type t_pv = [

| `Parameter of Signature.t * Names.ModuleName.t

]

val equal : t -> t -> bool

val hash : t -> int

val compare : t -> t -> int

val functor_arg_pos : t -> int

Gets the index in which the functor argument is, in the argument list. Useful to turn identifiers into unique anchors, since multiple arguments can have the same name.