Module Odoc_model__Names.ModuleTypeName

type t

val to_string : t -> string

val to_string_unsafe : t -> string

to_string_unsafe will allow even internal names to be turned into strings. Use with caution.

val of_string : string -> t

val of_ident : Ident.t -> t

val internal_of_string : string -> t

val internal_of_ident : Ident.t -> t

val is_internal : t -> bool

val equal : t -> t -> bool

val is_hidden : t -> bool

Hidden names are those that contain a double underscore, e.g. Hidden__module