Module Odoc_model_desc.Type_desc

type t('a) =

| Record(list( field ('a))) : t( 'a)

| Variant( ( 'a => case) ) : t( 'a)

| Pair( t( 'a), t( 'b)) : t( ('a, 'b) )

| Triple( t( 'a), t( 'b), t( 'c)) : t( ('a, 'b, 'c ) )

| List( t( 'a)) : t(list( 'a))

| Option( t( 'a)) : t(option( 'a))

| To_string( ( 'a => string) ) : t( 'a)

| Indirect( ( 'a => 'b) , t( 'b)) : t( 'a)

;

Runtime representation of types. It is only useful for inspecting values, not types, eg. it is possible to write a generic printer but not to deserialize.

and field('a) =

| F(string, ( 'a => 'b) , t( 'b)) : field ('a)

;

and case =

| C(string, 'b, t( 'b)) : case

| C0(string) : case

;

let bool: t(bool);

let string: t(string);

let int: t(int);