Module Odoc_odoc

module Compile : sig ... end

module Depends : sig ... end

Computes the dependencies required for each step of the pipeline to work correctly on a given input.

module Fs : sig ... end

Utilities to manipulate files and paths.

module Html_fragment : sig ... end

Produces html fragment files from a mld file.

module Html_page : sig ... end

module Indexing : sig ... end

module Latex : sig ... end

module Man_page : sig ... end

module Odoc_file : sig ... end

Load and save .odoc and .odocl files.

module Odoc_link : sig ... end

module Or_error : sig ... end

module Rendering : sig ... end

module Resolver : sig ... end

Management of the documentation environment.

module Source_tree : sig ... end

module Support_files : sig ... end

Copies odoc's support files (default theme and JS files) to a specified location.

module Url : sig ... end