Module Odoc_odoc.Compile

type parent_cli_spec =

| CliParent of string

| CliPackage of string

| CliNoparent

val name_of_output : prefix:string -> Fs.File.t -> string

Compute the name of the page from the output file. Prefix is the prefix to remove from the filename.

val resolve_parent_page : Resolver.t -> string -> (Odoc_model.Paths.Identifier.ContainerPage.t * Odoc_model.Lang.Page.child list, [> Or_error.msg ]) Or_error.result

Parse and resolve a parent reference. Returns the identifier of the parent and its children as a list of reference.

val compile : resolver:Resolver.t -> parent_cli_spec:parent_cli_spec -> hidden:bool -> children:string list -> output:Fs.File.t -> warnings_options:Odoc_model.Error.warnings_options -> source:(Fpath.t * string list) option -> cmt_filename_opt:string option -> Fs.File.t -> (unit, [> Or_error.msg ]) Or_error.result

Produces .odoc files out of .cm{i,t,ti} or .mld files.