Module Odoc_odoc.Env

Management of the documentation environment.

This is the module which does the link between packages, directories and DocOck's needs.

module Accessible_paths : sig ... end

val lookup_page : Accessible_paths.t -> string -> Odoc_model.Root.t option

val fetch_page : Accessible_paths.t -> Odoc_model.Root.t -> (Page.t, [> `Msg of string ]) Result.result

type t = Odoc_xref2.Env.resolver

type builder

val create : ?⁠important_digests:bool -> directories:Fs.Directory.t list -> open_modules:string list -> builder

Prepare the environment for a given list of include directories

parameter important_digests

indicate whether digests should be compared when doc-ock tries to lookup or fetch a unit. It defaults to true.

val build : builder -> [ `Unit of Compilation_unit.t | `Page of Page.t ] -> t

Initialize the environment for the given unit.