Module Odoc_odoc.Root

val load : string -> Stdlib.in_channel -> (Odoc_model.Root.t, [> Or_error.msg ]) Or_error.result

load fn ic reads a t from ic. fn is the name of the file ic is "watching", and is used for error reporting.

val read : Fs.File.t -> (Odoc_model.Root.t, [> Or_error.msg ]) Or_error.result

read f opens f for reading and then calls load.

val save : Stdlib.out_channel -> Odoc_model.Root.t -> unit

save oc t marshalls t to oc.