Module Odoc_odoc__Env

Management of the documentation environment.

This is the module which does the link between packages, directories and DocOck's needs.

type t

type builder

val create : ?⁠important_digests:bool -> directories:Odoc_odoc.Fs.Directory.t list -> builder

Prepare the environment for a given list of include directories

parameter important_digests

indicate whether digests should be compared when doc-ock tries to lookup or fetch a unit. It defaults to true.

val build : builder -> [ `Unit of Odoc_odoc.Compilation_unit.t | `Page of Odoc_odoc.Page.t ] -> t

Initialize the environment for the given unit.

val resolver : t -> Odoc_xref.resolver

Get a resolver from an env

val expander : t -> Odoc_xref.expander

Get an expander from an env