Module Odoc_xref2.Subst

type t = Component.Substitution.t ;

let identity: t;

let unresolve_opaque_paths: t => t;

let path_invalidate_module: Ident.path_module => t => t;

let mto_invalidate_module: Ident.path_module => t => t;

let add_module: Ident.path_module => Cpath.module_ => Cpath.Resolved.module_ => t => t;

let add_module_type: Ident.module_type => Cpath.module_type => Cpath.Resolved.module_type => t => t;

let add_type: Ident.type_ => Cpath.type_ => Cpath.Resolved.type_ => t => t;

let add_class: Ident.class_ => Cpath.class_type => Cpath.Resolved.class_type => t => t;

let add_class_type: Ident.class_type => Cpath.class_type => Cpath.Resolved.class_type => t => t;

let add_type_replacement: Ident.path_type => Component.TypeExpr.t => Component.TypeDecl.Equation.t => t => t;

let add_module_substitution: Ident.path_module => t => t;

let type_: t => Component.TypeDecl.t => Component.TypeDecl.t ;

let type_expr: t => Component.TypeExpr.t => Component.TypeExpr.t ;

let module_: t => Component.Module.t => Component.Module.t ;

let module_type: t => Component.ModuleType.t => Component.ModuleType.t ;

let module_substitution: t => Component.ModuleSubstitution.t => Component.ModuleSubstitution.t ;

let module_type_expr: t => Component.ModuleType.expr => Component.ModuleType.expr ;

let exception_: t => Component.Exception.t => Component.Exception.t ;

let extension: t => Component.Extension.t => Component.Extension.t ;

let external_: t => Component.External.t => Component.External.t ;

let include_: t => Component.Include.t => Component.Include.t ;

let open_: t => Component.Open.t => Component.Open.t ;

let value: t => Component.Value.t => Component.Value.t ;

let class_: t => Component.Class.t => Component.Class.t ;

let class_decl: t => Component.Class.decl => Component.Class.decl ;

let class_type: t => Component.ClassType.t => Component.ClassType.t ;

let signature: t => Component.Signature.t => Component.Signature.t ;

let apply_sig_map: t => list( Component.Signature.item ) => list( Component.Signature.removed_item ) => (list( Component.Signature.item ), list( Component.Signature.removed_item ), bool) ;

Apply substitutions. The third value is false if the corresponding signature needs to be compiled again and true otherwise.