Module Odoc_xref2__Tools

Tools for manipulating the component data structures

This module contains tools for manipulating the Component data structures, for example, resolving paths and fragments, obtaining signatures, handling fragment substitution and others.