Module Odoc_xref__.Resolve

type resolver

val build_resolver : ?⁠equal:(Odoc_model.Root.t -> Odoc_model.Root.t -> bool) -> ?⁠hash:(Odoc_model.Root.t -> int) -> (string -> Odoc_xref__.Component_table.lookup_unit_result) -> (Odoc_model.Root.t -> Odoc_model.Lang.Compilation_unit.t) -> (string -> Odoc_model.Root.t option) -> (Odoc_model.Root.t -> Odoc_model.Lang.Page.t) -> resolver

Lazily extract the components of units. Assumes that it is safe to use Hashtbl.hash and structural equality (=) on 'a.

val resolve : resolver -> Odoc_model.Lang.Compilation_unit.t -> Odoc_model.Lang.Compilation_unit.t

Try to resolve all paths and references within a unit.

val resolve_page : resolver -> Odoc_model.Lang.Page.t -> Odoc_model.Lang.Page.t

Try to resolve all paths and references within a page.