The odoc suite

The odoc distribution contains a set of packages that can be used to build documentation sites for OCaml projects. The individual projects are as follows:

odoc - the documentation tool itself

Sherlodoc - the recommended search tool

odoc-driver - a example driver for odoc

odoc-parser - a library to parse odoc markup

odoc-md - a tool to help integrate markdown pages into the odoc world