Mtime v2.0.0

Mtime has platform independent support for monotonic wall-clock time. This time increases monotonically and is not subject to operating system calendar time adjustments. The library has types to represent nanosecond precision timestamps and time spans.

Mtime_clock provides acces to a monotonic system clock and its resolution (if available).

Library mtime

Mtime Monotonic time values.

Library mtime.clock.os

Mtime_clock Monotonic time clock.

Also use this library for compiling to JavaScript.