Odig v0.0.9

Odig is a command line tool to lookup documentation of installed OCaml packages.

Consult the manual or the packaging conventions.

Quick start

A few commands to get you started:

odig doc           # Show API docs and manuals of installed packages
odig readme odig   # Consult the readme of odig
odig changes odig  # Consult the changelog of odig
odig browse issues odig  # Browse odig's issues.

API

This is an unstable API subject to change even between minor versions of the tool. Use at your own risk.

Odig_support Odig support library.

Odig_odoc odoc API reference generation.

Odig_odoc_page mld and HTML pages generated by odig itself.