Library opam-core

OpamACL C auxiliary function used for POSIX 1003.1e DRAFT 17 permission checking.

OpamCached

OpamCompat

OpamConsole Console output, ANSI color, logging and user querying

OpamCoreConfig Configuration options for the core lib (record, global reference and setter)

OpamCoreConfigDeveloper

OpamDirTrack

OpamFilename Higher level file and directory name manipulation AND file operations, wrappers on OpamSystem using the filename type

OpamHash

OpamJson Json encoder; only needed for some debug options

OpamParallel Parallel execution of jobs following a directed graph

OpamProcess Process and job handling, with logs, termination status, etc.

OpamSHA Pure OCaml implementation of SHA256/512 hashing functions. The hash is returned as an hex string.

OpamSWHID

OpamStd Generic stdlib functions (String, List, Option, Sys submodules...)

OpamStubs OS-specific functions requiring C code on at least one platform.

OpamStubsTypes Types for C stubs modules.

OpamSystem Bindings of lots of filesystem and system operations

OpamUrl URL parsing and printing, with support for our different backends

OpamVersion (generated) Current OPAM version

OpamVersionCompare Version comparison function used throughout. From the Dose suite.

OpamVersionInfo