opam-file-format

API

Library opam-file-format

OpamBaseParser

OpamLexer opam format lexer

OpamParser opam format parser

OpamParserTypes Defines the types for the opam format lexer and parser

OpamPrinter Functions for converting parsed opam files back to strings