ctypes

API

Library ctypes

ComplexL

Cstubs_internals

Ctypes The core ctypes module.

Ctypes_bigarray

Ctypes_bigarray_stubs

Ctypes_coerce

Ctypes_memory

Ctypes_memory_stubs

Ctypes_primitive_types

Ctypes_primitives

Ctypes_ptr

Ctypes_roots_stubs

Ctypes_static

Ctypes_std_view_stubs

Ctypes_std_views

Ctypes_structs

Ctypes_structs_computed Structs and unions whose layouts are computed from the sizes and alignment requirements of the constituent field types.

Ctypes_type_printing

Ctypes_types

Ctypes_value_printing

Ctypes_value_printing_stubs

LDouble

PosixTypes Some POSIX types.

Library ctypes.stubs

Cstubs Operations for generating C bindings stubs.

Cstubs_analysis

Cstubs_c_language

Cstubs_emit_c

Cstubs_errors

Cstubs_generate_c

Cstubs_generate_ml

Cstubs_inverted Operations for exposing OCaml code as C libraries.

Cstubs_public_name

Cstubs_structs

Ctypes_path

Library ctypes.top

Ctypes_top