Deps

The following libraries are dependencies of odoc.

Stdlib

Stdlib The OCaml Standard library.

Fpath

Fpath File system paths, file extensions, path sets and maps.

Cmdliner

Cmdliner Declarative definition of command line interfaces.

Astring

Astring Alternative Char and String modules.

Result

Result Result values.

Yojson

Yojson The Yojson library provides several types for representing JSON values, with different use cases.

Tyxml

Tyxml Typed implementation for HTML, SVG and XML

Biniou

Bi_dump

Bi_inbuf

Bi_io

Bi_outbuf

Bi_share

Bi_stream

Bi_util

Bi_vint

Odoc-parser

Odoc_parser Parser for ocamldoc formatted comments.

Fmt

Fmt Format pretty-printer combinators.

Fmt_cli Cmdliner support for Fmt.

Fmt_tty Fmt TTY setup.