Library compiler-libs.common

Annot

Arg_helper Decipher command line arguments of the form <value> | <key>=<value>,...

Ast_helper Helpers to produce Parsetree fragments

Ast_invariants Check AST invariants

Ast_iterator Ast_iterator.iterator enables AST inspection using open recursion. A typical mapper would be based on Ast_iterator.default_iterator, a trivial iterator, and will fall back on it for handling the syntax it does not modify.

Ast_mapper The interface of a -ppx rewriter

Asttypes Auxiliary AST types used by parsetree and typedtree.

Attr_helper Helpers for attributes

Binutils

Btype

Build_path_prefix_map Rewrite paths for reproducible builds

Builtin_attributes Support for the builtin attributes:

Bytesections

CamlinternalMenhirLib

Ccomp Compiling C files and building C libraries

Clflags Command line flags

Cmi_format

Cmo_format

Cmt2annot

Cmt_format cmt and cmti files format.

Compenv

Compile_common Common compilation pipeline between bytecode and native.

Compmisc

Compression

Config System configuration

Config_boot System configuration

Config_main System configuration

Consistbl Consistency tables: for checking consistency of module CRCs

Ctype

Datarepr

Debuginfo

Depend Module dependencies.

Diffing Parametric diffing

Diffing_with_keys When diffing lists where each element has a distinct key, we can refine the diffing patch by introducing two composite edit moves: swaps and moves.

Dll

Docstrings Documentation comments

Domainstate

Env

Envaux

Errortrace

Expunge

Ident

Identifiable Uniform interface for common data structures over various things.

Includeclass

Includecore

Includemod

Includemod_errorprinter

Int_replace_polymorphic_compare

Lambda

Lazy_backtrack

Lexer The lexical analyzer

Load_path Management of include directories.

Local_store This module provides some facilities for creating references (and hash tables) which can easily be snapshoted and restored to an arbitrary version.

Location Source code locations (ranges of positions), used in parsetree.

Longident Long identifiers, used in parsetree.

Main_args

Makedepend

Matching

Meta

Misc Miscellaneous useful types and functions

Mtype

Numbers Modules about numbers, some of which satisfy Identifiable.S.

Opcodes

Oprint

Outcometree

Parmatch Detection of partial matches and unused match cases.

Parse Entry points in the parser

Parser

Parsetree Abstract syntax tree produced by parsing

Path

Patterns

Persistent_env

Pparse Driver for the parser and external preprocessors.

Pprintast Pretty-printers for Parsetree

Predef

Primitive

Printast Raw printer for Parsetree

Printlambda

Printpat

Printtyp

Printtyped

Profile Compiler performance recording

Runtimedef

Shape Shapes are an abstract representation of modules' implementations which allow the tracking of definitions through functor applications and other module-level operations.

Shape_reduce

Signature_group Iterate on signature by syntactic group of items

Simplif Lambda simplification.

Strongly_connected_components Kosaraju's algorithm for strongly connected components.

Stypes

Subst

Switch

Symtable

Syntaxerr Auxiliary type for reporting syntax errors

Targetint Target processor-native integers.

Tast_iterator Allows the implementation of typed tree inspection using open recursion

Tast_mapper

Terminfo Basic interface to the terminfo database

Tmc Tail-modulo-cons optimization.

Translattribute

Translclass

Translcore

Translmod

Translobj

Translprim

Type_immediacy Immediacy status of a type

Typeclass

Typecore

Typedecl

Typedecl_immediacy

Typedecl_properties

Typedecl_separability The OCaml runtime assumes for type-directed optimizations that all types are "separable". A type is "separable" if either all its inhabitants (the values of this type) are floating-point numbers, or none of them are.

Typedecl_unboxed

Typedecl_variance

Typedtree Abstract syntax tree after typing

Typemod Type-checking of the module language and typed ast hooks

Typeopt

Types

Typetexp

Unit_info This module centralize the handling of compilation files and their metadata.

Untypeast

Value_rec_check

Value_rec_compiler

Value_rec_types Types related to the compilation of value let-recs (non-functional recursive definitions)

Warnings Warning definitions

X86_ast Structured representation of Intel assembly language (32 and 64 bit).