Library opam-format

OpamFile Handles all OPAM file formats as record types and submodules, conversion to and from syntax

OpamFilter Formulas on variables, as used in opam files build scripts

OpamFormat OPAM files syntax and conversion tools

OpamFormatConfig

OpamFormula Formulas on packages, opt. with sub-formulas on versions, and conversion functions

OpamInterpLexer OPAM format variable interpolation processor

OpamLineLexer A simple lexer to list of lines, which are lists of words

OpamPackage The package type, and package name type (name+version, values often called "nv" in the code)

OpamPath Defines the file hierarchy in ~/.opam

OpamPp Generic bidirectional transformation toolbox for parsing/printing

OpamRepositoryName The type for repository names

OpamSwitch The type for switch names

OpamSysPkg

OpamTypes Definitions of many types used throughout

OpamTypesBase Helper functions on the base types (from OpamTypes)

OpamVariable OPAM variables with scope (global or module), used in "opam" package definition files in "filters"