progress

API

Library progress

Progress A library for displaying progress bars, including support for rendering multiple bars at once. Start by describing of a sequence of progress bars, then begin rendering them.

Library progress.engine

Progress_engine Progress_engine provides a platform-independent progress bar renderer. Most users should use the Progress module, which instantiates this core with a Unix platform implementation.