terminal

API

Library terminal

Terminal This library provides a small set of standard utility functions for interacting with terminals.

Library terminal.ansi

Terminal_ansi This library provides a small set of standard utility functions for interacting with terminals.