Module Include2

module X : sig ... end

Comment about X that should not appear when including X below.

include module type of struct include X end

Comment about X that should not appear when including X below.

type t = int