Module Stop_dead_link_doc

type ('perm, 'record, 'field) t_with_perm =

| Field of ('perm, 'record, 'field) For_generated_code.t

'record is the type of the record. 'field is the type of the values stored in the record field with name name. 'perm is a way of restricting the operations that can be used.