Module Ocamlary

This is an interface with all of the module system features. C This text is centered. L This text is left-aligned. R This text is right-aligned. This documentation demonstrates:

comment formatting

unassociated comments

documentation sections

module system documentation including

submodules

module aliases

module types

module type aliases

modules with signatures

modules with aliased signatures

A numbered list:

3

2

1

David Sheets is the author.

author

David Sheets

Level 1

Level 2

Level 3

Level 4

Basic module stuff

module Empty: { ... };

A plain, empty module

module type Empty = { ... };

An ambiguous, misnamed module type

module type MissingComment = { ... };

An ambiguous, misnamed module type

Level 9000

module EmptyAlias = Empty;

A plain module alias of Empty

EmptySig

module type EmptySig = { ... };

A plain, empty module signature

module type EmptySigAlias = EmptySig;

A plain, empty module signature alias of

module ModuleWithSignature: EmptySig;

A plain module of a signature of EmptySig (reference)

module ModuleWithSignatureAlias: EmptySigAlias;

A plain module with an alias signature

module One: { ... };

module type SigForMod = { ... };

There's a signature in a module in this signature.

module type SuperSig = { ... };

module Buffer: { ... };

Buffer.t

Basic exception stuff

After exception title.

exception Kaboom(unit);

Unary exception constructor

exception Kablam(unit, unit);

Binary exception constructor

exception Kapow((unit, unit));

Unary exception constructor over binary tuple

exception EmptySig;

EmptySig is general but EmptySig is a module and EmptySig is this exception.

exception EmptySigAlias;

EmptySigAlias is this exception.

Basic type and value stuff with advanced doc comments

type a_function('a, 'b) = 'a => 'b;

a_function is general but a_function is this type and a_function is the value below.

let a_function: x:int => int;

This is a_function with param and return type.

parameter x

the x coordinate

returns

the y coordinate

let fun_fun_fun: a_function(a_function(int, int), a_function(unit, unit));

let fun_maybe: ?⁠yes:unit => unit => int;

let not_found: unit => unit;

raises Not_found

That's all it does

let ocaml_org: string;

see http://ocaml.org/

The OCaml Web site

let some_file: string;

see some_file

The file called some_file

let some_doc: string;

see some_doc

The document called some_doc

let since_mesozoic: unit;

This value was introduced in the Mesozoic era.

since

mesozoic

let changing: unit;

This value has had changes in 1.0.0, 1.1.0, and 1.2.0.

before 1.0.0

before 1.0.0

before 1.1.0

before 1.1.0

version

1.2.0

let with_foo: unit;

This value has a custom tag foo. @foo the body of the custom foo tag

Some Operators

let (~-): unit;

let (!): unit;

let (@): unit;

let ($): unit;

let (%): unit;

let (^): unit;

let (&): unit;

let (*): unit;

let (-): unit;

let (+): unit;

let (<): unit;

let (>): unit;

let (-?): unit;

let (/): unit;

let (-|): unit;

let (:=): unit;

let (=): unit;

let (land): unit;

Advanced Module Stuff

module CollectionModule: { ... };

This comment is for CollectionModule.

module type COLLECTION = module type of CollectionModule;

module type of

module Recollection: (C: COLLECTION) => COLLECTION with type collection = list(C.element) and type element = C.collection;

module type MMM = { ... };

module type RECOLLECTION = MMM with module C = Recollection(CollectionModule);

module type RecollectionModule = { ... };

module type A = { ... };

module type B = { ... };

module type C = { ... };

This module type includes two signatures.

module FunctorTypeOf: (Collection: module type of CollectionModule) => { ... };

This comment is for FunctorTypeOf.

module type IncludeModuleType = { ... };

This comment is for IncludeModuleType.

module type ToInclude = { ... };

include ToInclude;

module IncludedA: { ... };

module type IncludedB = { ... };

Advanced Type Stuff

type record = {

field1: int,
This comment is for field1.

field2: int,
This comment is for field2.

};

This comment is for record.

This comment is also for record.

type mutable_record = {

mutable a: int,
a is first and mutable

b: unit,
b is second and immutable

mutable c: int,
c is third and mutable

};

type universe_record = {

nihilate: a. 'a => unit,

};

type variant =

| TagA
This comment is for TagA.

| ConstrB(int)
This comment is for ConstrB.

| ConstrC(int, int)
This comment is for binary ConstrC.

| ConstrD((int, int))
This comment is for unary ConstrD of binary tuple.

;

This comment is for variant.

This comment is also for variant.

type poly_variant = [

| `TagA
This comment is for `TagA.

| `ConstrB(int)
This comment is for `ConstrB.

];

This comment is for poly_variant.

Wow! It was a polymorphic variant!

type full_gadt(_, _) =

| Tag : full_gadt(unit, unit)

| First('a) : full_gadt('a, unit)

| Second('a) : full_gadt(unit, 'a)

| Exist('a, 'b) : full_gadt('b, unit)

;

This comment is for full_gadt.

Wow! It was a GADT!

type partial_gadt('a) =

| AscribeTag : partial_gadt('a)

| OfTag(partial_gadt('a))

| ExistGadtTag(('a => 'b)) : partial_gadt('a)

;

This comment is for partial_gadt.

Wow! It was a mixed GADT!

type record_arg_gadt(_) =

| With_rec: {

foo: int,

} : record_arg_gadt(unit)

| With_poly_rec: {

bar: a. 'a => 'a,

} : record_arg_gadt(('a => 'a))

;

This comment is for record_arg_gadt.

Wow! It was a GADT with record arguments

type alias = variant;

This comment is for alias.

type tuple = ((alias, alias), alias, (alias, alias));

This comment is for tuple.

type variant_alias = variant =

| TagA

| ConstrB(int)

| ConstrC(int, int)

| ConstrD((int, int))

;

This comment is for variant_alias.

type record_alias = record = {

field1: int,

field2: int,

};

This comment is for record_alias.

type poly_variant_union = [

| poly_variant

| `TagC

];

This comment is for poly_variant_union.

type poly_poly_variant('a) = [

| `TagA('a)

];

type bin_poly_poly_variant('a, 'b) = [

| `TagA('a)

| `ConstrB('b)

];

type open_poly_variant('a) = [> `TagA ] as a;

type open_poly_variant2('a) = [> `ConstrB(int) ] as a;

type open_poly_variant_alias('a) = open_poly_variant2(open_poly_variant('a));

type poly_fun('a) = [> `ConstrB(int) ] as a => 'a;

type poly_fun_constraint('a) = 'a => 'a constraint 'a = [> `TagA ];

type closed_poly_variant('a) = [< `One | `Two ] as a;

type clopen_poly_variant('a) = [< `One | `Two(int) | `Three Two Three ] as a;

type nested_poly_variant = [

| `A

| `B([ `B1 | `B2 ])

| `C

| `D([ `D1([ `D1a ]) ])

];

type full_gadt_alias('a, 'b) = full_gadt('a, 'b) =

| Tag : full_gadt_alias(unit, unit)

| First('a) : full_gadt_alias('a, unit)

| Second('a) : full_gadt_alias(unit, 'a)

| Exist('a, 'b) : full_gadt_alias('b, unit)

;

This comment is for full_gadt_alias.

type partial_gadt_alias('a) = partial_gadt('a) =

| AscribeTag : partial_gadt_alias('a)

| OfTag(partial_gadt_alias('a))

| ExistGadtTag(('a => 'b)) : partial_gadt_alias('a)

;

This comment is for partial_gadt_alias.

exception Exn_arrow(unit) : exn;

This comment is for Exn_arrow.

type mutual_constr_a =

| A

| B_ish(mutual_constr_b)
This comment is between mutual_constr_a and mutual_constr_b.

;

This comment is for mutual_constr_a then mutual_constr_b.

and mutual_constr_b =

| B

| A_ish(mutual_constr_a)
This comment must be here for the next to associate correctly.

;

This comment is for mutual_constr_b then mutual_constr_a.

type rec_obj = {. f: int, g: unit => unit, h: rec_obj, };

type open_obj('a) = {.. f: int, g: unit => unit, } as a;

type oof('a) = {.. a: unit, } as a => 'a;

type any_obj('a) = {.. } as a;

type empty_obj = {. };

type one_meth = {. meth: unit, };

type ext = ..;

A mystery wrapped in an ellipsis

type ext += ExtA;

type ext += ExtB;

type ext += ExtC(unit) | ExtD(ext);

type ext += ExtE;

type ext += ExtF;

type poly_ext('a) = ..;

'a poly_ext

type poly_ext += Foo('b) | Bar('b, 'b);

type poly_ext += Quux('c);

module ExtMod: { ... };

type ExtMod.t += ZzzTop0;

type ExtMod.t += ZzzTop(unit);

let launch_missiles: unit => unit;

Rotate keys on my mark...

type my_mod = (module COLLECTION);

A brown paper package tied up with string

class empty_class: { ... }

class one_method_class: { ... }

class two_method_class: { ... }

class ('a) param_class: 'a => { ... }

type my_unit_object = param_class(unit);

type my_unit_class('a) = param_class(unit) as a;

module Dep1: { ... };

module Dep2: (Arg: { ... }) => { ... };

type dep1 = Dep2(Dep1).B.c;

module Dep3: { ... };

module Dep4: { ... };

module Dep5: (Arg: { ... }) => { ... };

type dep2 = Dep5(Dep4).Z.X.b;

type dep3 = Dep5(Dep4).Z.Y.a;

module Dep6: { ... };

module Dep7: (Arg: { ... }) => { ... };

type dep4 = Dep7(Dep6).M.Y.d;

module Dep8: { ... };

module Dep9: (X: { ... }) => { ... };

module type Dep10 = Dep9(Dep8).T with type t = int;

module Dep11: { ... };

module Dep12: (Arg: { ... }) => { ... };

module Dep13: Dep12(Dep11).T;

type dep5 = Dep13.c;

module type With1 = { ... };

module With2: { ... };

module With3: With1 with module M = With2;

type with1 = With3.N.t;

module With4: With1 with module M := With2;

type with2 = With4.N.t;

module With5: { ... };

module With6: { ... };

module With7: (X: { ... }) => { ... };

module type With8 = With7(With6).T with module M = With5 and type M.N.t = With5.N.t;

module With9: { ... };

module With10: { ... };

module type With11 = With7(With10).T with module M = With9 and type N.t = int;

module type NestedInclude1 = { ... };

include NestedInclude1;

module type NestedInclude2 = { ... };

include NestedInclude2 with type nested_include = int;

type nested_include = int;

module DoubleInclude1: { ... };

module DoubleInclude3: { ... };

include module type of DoubleInclude3.DoubleInclude2;

type double_include;

module IncludeInclude1: { ... };

include module type of IncludeInclude1;

module type IncludeInclude2 = { ... };

include IncludeInclude2;

type include_include;

Trying the {!modules: ...} command.

With ocamldoc, toplevel units will be linked and documented, while submodules will behave as simple references.

With odoc, everything should be resolved (and linked) but only toplevel units will be documented.

Dep1.X

DocOckTypes

Ocamlary.IncludeInclude1

Ocamlary

Weirder usages involving module types

IncludeInclude1.IncludeInclude2

Dep4.T

A.Q

Playing with @canonical paths

module CanonicalTest: { ... };

let test: CanonicalTest.Base.List.t('a) => unit;

Some ref to CanonicalTest.Base__Tests.C.t and CanonicalTest.Base__Tests.L.id. But also to Ocamlary.CanonicalTest.Base.List and Ocamlary.CanonicalTest.Base.List.t

Aliases again

module Aliases: { ... };

Section title splicing

I can refer to

{!section:indexmodules} : Trying the {!modules: ...} command.

{!aliases} : Aliases again

But also to things in submodules:

{!section:SuperSig.SubSigA.subSig} : SuperSig.SubSigA:subSig

{!Aliases.incl} : Aliases:incl

And just to make sure we do not mess up:

{{!section:indexmodules}A} : A

{{!aliases}B} : B

{{!section:SuperSig.SubSigA.subSig}C} : C

{{!Aliases.incl}D} : D

New reference syntax

module type M = { ... };

module M: { ... };

module Only_a_module: { ... };