Module Stop_dead_link_doc

type t_with_perm('perm, 'record, 'field) =

| Field(For_generated_code.t('perm, 'record, 'field))

;

'record is the type of the record. 'field is the type of the values stored in the record field with name name. 'perm is a way of restricting the operations that can be used.