Uucp v16.0.0

Uucp provides efficient access to a selection of character properties of the Unicode character database.

Consult this page for a minimal Unicode introduction and OCaml Unicode tips.

Library uucp

Uucp Unicode character properties.