odoc

What is odoc?

odoc is a documentation generator for OCaml. It reads doc comments, delimited with (** ... *), and outputs HTML, LaTeX and man pages. The pages you are reading now are rendered using odoc.

Text inside doc comments is marked up in ocamldoc syntax:

val compare : string -> string -> int
(** [compare s1 s2] compares [s1] and [s2] in {e lexicographic} order. The
    result is negative if [s1] precedes [s2], positive if [s1] follows [s2],
    and zero if [s1] and [s2] are equal. *)

The syntax reference is a refinement of that explained in the OCaml manual. The differences are described here.

odoc's main advantage over ocamldoc is an accurate cross-referencer, which handles the complexity of the OCaml module system, and an expander, which can expand complex module-type expressions while preserving their structure, including comment, includes and more.

For library authors

If you are looking for guidance on how to document your OCaml project, see odoc for authors.

For integrators

If you are looking to integrate odoc into your tool, webpage or any other setting, you will need to understand how to drive odoc.

For developers

If you want to start contributing to odoc, please see our contributors guide.