Module Odoc_document.Codefmt

type out = Types.Source.t

module State : sig ... end

Modern implementation using semantic tags, Only for 4.08+

module Tag : sig ... end

Ugly terrible implementation of Format Semantic tags for OCaml < 4.08. Please get rid of it as soon as possible.

val make : unit -> (unit -> Types.Source.token list) * Format.formatter

val spf : ('a, Format.formatter, unit, Types.Source.token list) format4 -> 'a

val pf : Format.formatter -> ('a, Format.formatter, unit) format -> 'a

Transitory hackish API

val elt : Format.formatter -> Types.Inline.t -> unit

val entity : Types.Inline.entity -> Format.formatter -> unit

val (++) : ('a -> 'b) -> ('a -> 'c) -> 'a -> 'c

val span : ?attr:string -> (Format.formatter -> unit) -> Format.formatter -> unit

val txt : string -> Format.formatter -> unit

val noop : Format.formatter -> unit

val break : int -> int -> Format.formatter -> unit

val cut : Format.formatter -> unit

val sp : Format.formatter -> unit

val (!) : 'a Fmt.t -> 'a -> Format.formatter -> unit

val list : ?sep:(Format.formatter -> 'a) -> f:('b -> Format.formatter -> unit) -> 'b list -> Format.formatter -> unit

val render : (Format.formatter -> unit) -> Types.Source.token list

val code : ?attr:Types.Class.t -> (Format.formatter -> unit) -> Types.Inline.one list

val documentedSrc : (Format.formatter -> unit) -> Types.DocumentedSrc.one list

val codeblock : ?attr:Types.Class.t -> (Format.formatter -> unit) -> Types.Block.one list

val keyword : string -> Format.formatter -> unit

module Infix : sig ... end