Module Odoc_document.Codefmt

type out = Types.Source.t ;

module State : { ... };

Modern implementation using semantic tags, Only for 4.08+

module Tag : { ... };

Ugly terrible implementation of Format Semantic tags for OCaml < 4.08. Please get rid of it as soon as possible.

let make: unit => ( (unit => list( Types.Source.token )) , Stdlib.Format.formatter ) ;

let spf: Stdlib.format4 ('a, Stdlib.Format.formatter, unit , list( Types.Source.token )) => 'a;

let pf: Stdlib.Format.formatter => Stdlib.format ('a, Stdlib.Format.formatter, unit) => 'a;

Transitory hackish API

let elt: Stdlib.Format.formatter => Types.Inline.t => unit;

let entity: Types.Inline.entity => Stdlib.Format.formatter => unit;

let (++): ( 'a => 'b) => ( 'a => 'c) => 'a => 'c;

let span: ?attr:string => ( Stdlib.Format.formatter => unit) => Stdlib.Format.formatter => unit;

let txt: string => Stdlib.Format.formatter => unit;

let noop: Stdlib.Format.formatter => unit;

let break: int => int => Stdlib.Format.formatter => unit;

let cut: Stdlib.Format.formatter => unit;

let sp: Stdlib.Format.formatter => unit;

let (!): Fmt.t( 'a) => 'a => Stdlib.Format.formatter => unit;

let list: ?sep: ( Stdlib.Format.formatter => 'a) => f: ( 'b => Stdlib.Format.formatter => unit) => list('b) => Stdlib.Format.formatter => unit;

let render: ( Stdlib.Format.formatter => unit) => list( Types.Source.token );

let code: ?attr: Types.Class.t => ( Stdlib.Format.formatter => unit) => list( Types.Inline.one );

let documentedSrc: ( Stdlib.Format.formatter => unit) => list( Types.DocumentedSrc.one );

let codeblock: ?attr: Types.Class.t => ( Stdlib.Format.formatter => unit) => list( Types.Block.one );

let keyword: string => Stdlib.Format.formatter => unit;

module Infix : { ... };