Module Odoc_document.Codefmt

type out = Types.Source.t

module State : sig ... end

module Tag : sig ... end

Ugly terrible implementation of Format Semantic tags for OCaml < 4.08. Please get rid of it as soon as possible.

val make : unit -> (unit -> Types.Source.token list) * Stdlib.Format.formatter

val spf : ('a, Stdlib.Format.formatter, unit, Types.Source.token list) Stdlib.format4 -> 'a

val pf : Stdlib.Format.formatter -> ('a, Stdlib.Format.formatter, unit) Stdlib.format -> 'a

val elt : Stdlib.Format.formatter -> Types.Inline.t -> unit

val entity : Types.Inline.entity -> Stdlib.Format.formatter -> unit

val (++) : ('a -> 'b) -> ('a -> 'c) -> 'a -> 'c

val span : (Stdlib.Format.formatter -> unit) -> Stdlib.Format.formatter -> unit

val txt : string -> Stdlib.Format.formatter -> unit

val noop : Stdlib.Format.formatter -> unit

val (!) : 'a Fmt.t -> 'a -> Stdlib.Format.formatter -> unit

val list : ?⁠sep:(Stdlib.Format.formatter -> 'a) -> f:('b -> Stdlib.Format.formatter -> unit) -> 'b list -> Stdlib.Format.formatter -> unit

val render : (Stdlib.Format.formatter -> unit) -> Types.Source.token list

val code : ?⁠attr:Types.Class.t -> (Stdlib.Format.formatter -> unit) -> Types.Inline.one list

val documentedSrc : (Stdlib.Format.formatter -> unit) -> Types.DocumentedSrc.one list

val codeblock : ?⁠attr:Types.Class.t -> (Stdlib.Format.formatter -> unit) -> Types.Block.one list

val keyword : string -> Stdlib.Format.formatter -> unit

module Infix : sig ... end