Module Codefmt.Tag

Ugly terrible implementation of Format Semantic tags for OCaml < 4.08. Please get rid of it as soon as possible.

val setup_tags : Stdlib.Format.formatter -> State.t -> unit

val elt : Stdlib.Format.formatter -> Types.Inline.t -> unit