Module Markup.Preamble_examples

This module demonstrates the various ways that preambles are calculated

module Module : sig ... end

This comment is on the declaration of Module as an alias of Hidden__Module

module Module2 : sig ... end

This is the top comment declared in the module Hidden__module2.

module Nonhidden_module : sig ... end

This is the top comment declared in the module Hidden__module2.

module Module3 = Nonhidden_module

This comment is on the declaration of Module3 as an alias of Nonhidden_module

module Nonhidden_module2 : sig ... end

This is the top comment declared in the module Hidden__module2.

module Module4 = Nonhidden_module2

The modules special reference can be used to refer to a list of modules. It uses the synopsis from the modules