Module Odoc_examples.Resolution

Examples of Path, Fragment and Reference Resolution

This module contains examples of some of the features of Resolution as described in the page features. See the explanations there for details on what each of these demonstrates.

module Alias : sig ... end

Demonstrates a reference to an item in a module that's an alias

module HiddenAlias : sig ... end

Demonstrates a reference to an item in a module that's an alias of a hidden module.

module Canonical : sig ... end

Demonstrates the use of canonical tags

module Fragments : sig ... end

Demonstrates the resolution of fragments

module Hidden : sig ... end

Demonstrates paths to hidden items

module References : sig ... end

Examples of resolution of references

module Complicated_1 : sig ... end

Some more complicated examples of resolution

module Complicated_2 : sig ... end

A very complicated example of resolution