Module Odoc_odoc.Env

Management of the documentation environment.

This is the module which does the link between packages, directories and DocOck's needs.

module Accessible_paths : { ... };

let lookup_page: Accessible_paths.t => string => option( Odoc_model.Root.t );

let fetch_page: Accessible_paths.t => Odoc_model.Root.t => Result.result ( Page.t , [> `Msg(string) ]) ;

type t = Odoc_xref2.Env.resolver ;

type builder ;

let create: ?important_digests:bool => directories:list( Fs.Directory.t ) => open_modules:list(string) => builder;

Prepare the environment for a given list of include directories

parameter important_digests
indicate whether digests should be compared when doc-ock tries to lookup or fetch a unit. It defaults to true.

let build: builder => [ `Unit ( Compilation_unit.t ) | `Page ( Page.t ) ] => t;

Initialize the environment for the given unit.