Module Odoc_odoc.Resolver

Management of the documentation environment.

This is the module which does the link between packages, directories and Odoc_xref2's needs.

type t

val create : important_digests:bool -> directories:Fs.Directory.t list -> open_modules:string list -> t

Prepare the environment for a given list of include directories

parameter important_digests
indicate whether digests should be compared when odoc_xref2 tries to lookup or fetch a unit. It defaults to true.

val lookup_page : t -> string -> Odoc_model.Lang.Page.t option

val build_env_for_unit : t -> Odoc_model.Lang.Compilation_unit.t -> Odoc_xref2.Env.t

Initialize the environment for the given module.

val build_env_for_page : t -> Odoc_model.Lang.Page.t -> Odoc_xref2.Env.t

Initialize the environment for the given page.

val resolve_import : t -> string -> Odoc_model.Root.t option

Similar to Odoc_xref2.Env.lookup_root_module but save work by loading only the root. Only used when resolving imports, which are needed for the link-deps command.