Module Odoc_xref2.Component

module ModuleMap : Stdlib.Map.S with type key = Ident.module_

module TypeMap : Stdlib.Map.S with type key = Ident.type_

module PathModuleMap : Stdlib.Map.S with type key = Ident.path_module

Useful maps

module ModuleTypeMap : Stdlib.Map.S with type key = Ident.module_type

module PathTypeMap : Stdlib.Map.S with type key = Ident.path_type

module PathClassTypeMap : Stdlib.Map.S with type key = Ident.path_class_type

module IdentMap : Stdlib.Map.S with type key = Ident.any

module Delayed : sig ... end

Delayed is a bit like Lazy.t but may in the future offer the chance to peek inside to be able to optimize the calculation

module Opt : sig ... end

Components

module rec Module : sig ... end

and ModuleSubstitution : sig ... end

and TypeExpr : sig ... end

and Extension : sig ... end

and Exception : sig ... end

and FunctorParameter : sig ... end

and ModuleType : sig ... end

and TypeDecl : sig ... end

and Value : sig ... end

and Signature : sig ... end

and Open : sig ... end

and Include : sig ... end

and External : sig ... end

and Class : sig ... end

and ClassType : sig ... end

and ClassSignature : sig ... end

and Method : sig ... end

and InstanceVariable : sig ... end

and Substitution : sig ... end

and CComment : sig ... end

module Element : sig ... end

module Fmt : sig ... end

Formatting functions for components

module Of_Lang : sig ... end

val module_of_functor_argument : FunctorParameter.parameter -> Module.t