Deps

The following libraries are dependencies of odoc.

Stdlib

Stdlib The standard library of OCaml.

Fpath

Fpath provides a library for handling file system paths.

Cmdliner

Cmdliner provides a library for handling command-line arguments.

Astring

Astring provides a replacement for the standard OCaml String module.

Result

Result is the Result Monad module.

Yojson

Yojson is a JSON library.

Tyxml

Tyxml is the library used to generate HTML pages.

Biniou

Bi_dump Bi_inbuf Bi_io Bi_outbuf Bi_share Bi_stream Bi_util Bi_vint - Binary data format designed for speed, safety, ease of use and backward compatibility as protocols evolve

Odoc-parser

Odoc_parser is the parser library used

Base

Base Base_internalhash_types Caml Md5_lib Shadow_stdlib

Base_quickcheck

Base_quickcheck Ppx_quickcheck Ppx_quickcheck_expander

Fmt

Fmt Fmt_cli Fmt_tty

Ppx_hash

Ppx_hash Ppx_hash_expander Ppx_hash_lib

Ppx_sexp_conv

Ppx_sexp_conv Ppx_sexp_conv_lib Ppx_sexp_conv_expander

Sexplib

Sexplib Sexplib_num Sexplib_unix

Sexplib0

Sexplib0

Core_kernel

core_kernel